requirements.txt
discord.py
requests
openai
aiohttp
duckduckgo_search
tenacity